Fetch current version

+++ tools/codegen/core/gen_upb_api.sh
@@ -14,33 +14,41 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# REQUIRES: Bazel
 set -ex
 
+cd $(dirname $0)/../../..
+bazel=`pwd`/tools/bazel
+
+if [ $# -eq 0 ]; then
+  UPB_OUTPUT_DIR=$PWD/src/core/ext/upb-generated
+  rm -rf $UPB_OUTPUT_DIR
+  mkdir -p $UPB_OUTPUT_DIR
+else
+  UPB_OUTPUT_DIR=$1
+fi
+
 pushd third_party/protobuf
-bazel build :protoc
+$bazel build :protoc
 PROTOC=$PWD/bazel-bin/protoc
 popd
 
 pushd third_party/upb
-bazel build :protoc-gen-upb
+$bazel build :protoc-gen-upb
 UPB_PLUGIN=$PWD/bazel-bin/protoc-gen-upb
 popd
-
-UPB_OUTPUT_DIR=$PWD/src/core/ext/upb-generated
-rm -rf $UPB_OUTPUT_DIR
-mkdir $UPB_OUTPUT_DIR
 
 proto_files=( \
   "envoy/api/v2/auth/cert.proto" \
   "envoy/api/v2/cds.proto" \
   "envoy/api/v2/cluster/circuit_breaker.proto" \
+  "envoy/api/v2/cluster/filter.proto" \
   "envoy/api/v2/cluster/outlier_detection.proto" \
   "envoy/api/v2/core/address.proto" \
   "envoy/api/v2/core/base.proto" \
   "envoy/api/v2/core/config_source.proto" \
   "envoy/api/v2/core/grpc_service.proto" \
   "envoy/api/v2/core/health_check.proto" \
+  "envoy/api/v2/core/http_uri.proto" \
   "envoy/api/v2/core/protocol.proto" \
   "envoy/api/v2/discovery.proto" \
   "envoy/api/v2/eds.proto" \
@@ -61,21 +69,23 @@ proto_files=( \
   "google/protobuf/timestamp.proto" \
   "google/protobuf/wrappers.proto" \
   "google/rpc/status.proto" \
-  "grpc/gcp/altscontext.proto" \
-  "grpc/gcp/handshaker.proto" \
-  "grpc/gcp/transport_security_common.proto" \
-  "grpc/health/v1/health.proto" \
-  "grpc/lb/v1/load_balancer.proto" \
+  "src/proto/grpc/gcp/altscontext.proto" \
+  "src/proto/grpc/gcp/handshaker.proto" \
+  "src/proto/grpc/gcp/transport_security_common.proto" \
+  "src/proto/grpc/health/v1/health.proto" \
+  "src/proto/grpc/lb/v1/load_balancer.proto" \
+  "udpa/data/orca/v1/orca_load_report.proto" \
   "validate/validate.proto")
 
 for i in "${proto_files[@]}"
 do
+  echo "Compiling: ${i}"
   $PROTOC \
+    -I=$PWD/third_party/udpa \
     -I=$PWD/third_party/envoy-api \
     -I=$PWD/third_party/googleapis \
     -I=$PWD/third_party/protobuf/src \
     -I=$PWD/third_party/protoc-gen-validate \
-    -I=$PWD/src/proto \
     -I=$PWD \
     $i \
     --upb_out=$UPB_OUTPUT_DIR \
